Converted the numbers to bytes so that it would compile.

+++ RawConnTest.go
@@ -49,9 +49,9 @@ func (c *UDP) read(size int) ([]byte, error) {
 }
 func (c *UDP) write(x []byte) error {
 	UDPHeader := []byte{
-		c.src >> 8, c.src << 8 >> 8, // Source port in byte slice
-		c.dest >> 8, c.dest << 8 >> 8, // Dest port in byte slice
-		(8 + len(x)) >> 8, (8 + len(x)) << 8 >> 8, // Length in bytes of UDP header + data
+        (byte)(c.src >> 8), (byte)(c.src << 8 >> 8), // Source port in byte slice
+        (byte)(c.dest >> 8), (byte)(c.dest << 8 >> 8), // Dest port in byte slice
+        (byte)((8 + len(x)) >> 8), (byte)((8 + len(x)) << 8 >> 8), // Length in bytes of UDP header + data
 		0, 0, // Checksum
 	}
